updated routes and modules

+++ routes/api/product-routes.js
@@ -5,22 +5,16 @@ const { Product, Category, Tag, ProductTag } = require('../../models');
 
 // get all products
 router.get('/', (req, res) => {
-  // find all products
-  // be sure to include its associated Category and Tag data
   Product.findAll({
-    attributes: ['id', 'product_name', 'price', 'stock'],
     include: [
-      {
-        model: Category,
-        attributes: ['category_name']
-      },
+      Category,
       {
         model: Tag,
-        attributes: ['tag_name']
+        through: ProductTag
       }
     ]
   })
-    .then(dbProductData => res.json(dbProductData))
+    .then(products => res.json(products))
     .catch(err => {
       console.log(err);
       res.status(500).json(err);
@@ -29,38 +23,26 @@ router.get('/', (req, res) => {
 
 // get one product
 router.get('/:id', (req, res) => {
-  // find a single product by its `id`
-  // be sure to include its associated Category and Tag data
   Product.findOne({
     where: {
       id: req.params.id
     },
-    attributes: ['id', 'product_name', 'price', 'stock'],
     include: [
-      {
-        model: Category,
-        attributes: ['category_name']
-      },
+      Category,
       {
         model: Tag,
-        attributes: ['tag_name']
+        through: ProductTag
       }
     ]
   })
-    .then(dbProductData => {
-      if (!dbProductData) {
-        res.status(404).json({message: 'No product was found with this id'});
-        return;
-      }
-      res.json(dbProductData);
-    })
+    .then(products => res.json(products))
     .catch(err => {
       console.log(err);
-      res.status(500).json(err);
+      res.status(400).json(err);
     });
 });
 
-// create new product
+// creat new product
 router.post('/', (req, res) => {
   /* req.body should look like this...
     {
@@ -70,29 +52,23 @@ router.post('/', (req, res) => {
       tagIds: [1, 2, 3, 4]
     }
   */
-  Product.create({
-    product_name: req.body.product_name,
-    price: req.body.price,
-    stock: req.body.stock,
-    category_id: req.body.category_id,
-    tagIds: req.body.tagIds
-    })
-    .then((product) => {
-      // If there are product tags, we need to create pairings in the ProductTag model to create in bulk.
-      if (req.body.tagIds.length) {
-        const productTagIdArr = req.body.tagIds.map((tag_id) => {
+  Product.create(req.body)
+    .then(product => {
+      // if there's product tags, we need to create pairings to bulk create in the ProductTag model
+      if (req.body.tagIds && req.body.tagIds.length) {
+        const productTagIdArr = req.body.tagIds.map(tag_id => {
           return {
             product_id: product.id,
-            tag_id,
+            tag_id
           };
         });
         return ProductTag.bulkCreate(productTagIdArr);
       }
-      // Simply respond if there are no product tags.
+      // if no product tags, just respond
       res.status(200).json(product);
     })
-    .then((productTagIds) => res.status(200).json(productTagIds))
-    .catch((err) => {
+    .then(productTagIds => res.status(200).json(productTagIds))
+    .catch(err => {
       console.log(err);
       res.status(400).json(err);
     });
@@ -100,64 +76,61 @@ router.post('/', (req, res) => {
 
 // update product
 router.put('/:id', (req, res) => {
-  // update information about the product
+  // update product data
   Product.update(req.body, {
     where: {
-      id: req.params.id,
-    },
+      id: req.params.id
+    }
   })
-    .then((ProductTag) => {
-      // locate all ProductTag-associated tags.
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // the current tag_ids list
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // make a list of new tag_ids in a filter.
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
+    .then(product => {
+      if (req.body.tagIds && req.body.tagIds.length) {
+        const productTags = ProductTag.findAll({
+          where: { product_id: req.params.id }
         });
-      // determine which ones to eliminate.
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+        const productTagIds = productTags.map(({ tag_id }) => tag_id);
+        // create filtered list of new tag_ids
+        const newProductTags = req.body.tagIds
+          .filter(tag_id => !productTagIds.includes(tag_id))
+          .map(tag_id => {
+            return {
+              product_id: req.params.id,
+              tag_id
+            };
+          });
+        // figure out which ones to remove
+        const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id);
 
-      // execute both steps
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+        // run both actions
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags)
+        ]);
+      }
+
+      return res.json(product);
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
+    .catch(err => {
       // console.log(err);
       res.status(400).json(err);
     });
 });
 
 router.delete('/:id', (req, res) => {
-  // delete one product by its `id` value
   Product.destroy({
     where: {
       id: req.params.id
     }
   })
-  .then(dbProductData => {
-    if (!dbProductData) {
-      rs.status(404).json({message: 'No product was found with this id'});
-      return;
-    }
-    res.json(dbProductData);
-  })
-  .catch(err => {
-    console.log(err);
-    res.status(500).json(err);
-  });
+    .then(products => {
+      console.log(products);
+      res.json(products);
+    })
+    .catch(err => {
+      console.log(err);
+      res.status(400).json(err);
+    });
 });
 
 module.exports = router;
